refact: 라우터 헤더부분 레이아웃과 outlet 을 사용하여 수정
