+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Header from "./components/Header";
+import Layout from "./components/layouts/Layouts";
 import Product from "./components/Products";
 import AddItem from "./components/AddItem";
 import GlobalStyle from "./css/GlobalStyle";
@@ -11,13 +11,14 @@ function App() {
   return (
     <Router>
       <GlobalStyle />
-      <Header />
       <Routes>
-        <Route path="/" element={<Product />} />
-        <Route path="/items" element={<Product />} />
-        <Route path="/additem" element={<AddItem />} />
+        <Route path="/" element={<Layout />}>
+          <Route index element={<Product />} />
+          <Route path="/items" element={<Product />} />
+          <Route path="/additem" element={<AddItem />} />
+          <Route path="/items/:productId" element={<ProductDetailPage />} />
+        </Route>
         <Route path="/login" element={<LoginPage />} />
-        <Route path="/items/:productId" element={<ProductDetailPage />} />
       </Routes>
     </Router>
   );
