+++ src/components/CommentList.js
@@ -5,6 +5,7 @@ import NoneImg from "../assets/images/Img_inquiry_empty.png";
 import BackIcon from "../assets/images/ic_back.png";
 import styled from "styled-components";
 import KebabIcon from "../assets/images/ic_kebab.png";
+import { formatTimes } from "../utils/FormatTimes";
 
 const CommentList = () => {
   const { productId } = useParams();
@@ -28,23 +29,6 @@ const CommentList = () => {
     fetchComments();
   }, [productId]);
 
-  function formatTimeSince(dateString) {
-    const now = new Date();
-    const updatedAt = new Date(dateString);
-    const diffInSeconds = Math.floor(
-      (now.getTime() - updatedAt.getTime()) / 1000
-    );
-
-    if (diffInSeconds < 60) {
-      return "방금 전";
-    } else if (diffInSeconds < 3600) {
-      return `${Math.floor(diffInSeconds / 60)}분 전`;
-    } else if (diffInSeconds < 86400) {
-      return `${Math.floor(diffInSeconds / 3600)}시간 전`;
-    } else {
-      return `${Math.floor(diffInSeconds / 86400)}일 전`;
-    }
-  }
   return (
     <>
       {!comments || comments.length === 0 ? (
@@ -73,7 +57,7 @@ const CommentList = () => {
                   />
                   <NicknameWrapper>
                     <NickName>{comment.writer.nickname}</NickName>
-                    <UpdateAt>{formatTimeSince(comment.updatedAt)}</UpdateAt>
+                    <UpdateAt>{formatTimes(comment.updatedAt)}</UpdateAt>
                   </NicknameWrapper>
                 </ProfileWrapper>
                 <Divider />
