+++ src/utils/FormatTimes.js
@@ -1,0 +1,17 @@
+export const formatTimes = (dateString) => {
+  const now = new Date();
+  const updatedAt = new Date(dateString);
+  const diffInSeconds = Math.floor(
+    (now.getTime() - updatedAt.getTime()) / 1000
+  );
+
+  if (diffInSeconds < 60) {
+    return "방금 전";
+  } else if (diffInSeconds < 3600) {
+    return `${Math.floor(diffInSeconds / 60)}분 전`;
+  } else if (diffInSeconds < 86400) {
+    return `${Math.floor(diffInSeconds / 3600)}시간 전`;
+  } else {
+    return `${Math.floor(diffInSeconds / 86400)}일 전`;
+  }
+};
